BigDecimal for money and 2 decimals for tableView salary content

+++ src/nl/inholland/model/Teacher.java
@@ -1,18 +1,19 @@
 package nl.inholland.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class Teacher extends User {
 
-    private final Double salary;
+    private final BigDecimal salary;
 
-    public Teacher(int id, String firstName, String lastName, String email, String password, LocalDate dateOfBirth, Double salary) {
+    public Teacher(int id, String firstName, String lastName, String email, String password, LocalDate dateOfBirth, BigDecimal salary) {
         super(id, firstName, lastName, email, password, dateOfBirth, UserType.EDITOR);
 
         this.salary = salary;
     }
 
-    public Double getSalary(){
+    public BigDecimal getSalary(){
         return salary;
     }
 }
